[Chore #2178134] Refactor test folder directory
